requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.12.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# HTTP requests
requests>=2.25.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0

# Optional: for production deployment
gunicorn>=20.1.0
